pyproject.toml
[project]
name = "wwpppp"
version = "0.1.0"
description = "Watcher for WPlace Paint Project Progress"
authors = [
    { name = "Liz Balbuena" }
]
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "platformdirs>=4.5.0",
    "requests>=2.32.5",
    "watchfiles>=1.1.1",
]

[project.scripts]
wwpppp = "wwpppp.main:main"

[tool.ruff]
line-length = 120

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
addopts = "--cov=wwpppp --cov-report=term-missing --cov-fail-under=95 -q"
python_files = ["tests/*.py"]
